% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCC_igraph.R
\name{SCC_igraph}
\alias{SCC_igraph}
\title{WeimingSCC::SCC_igraph}
\usage{
SCC_igraph(edge.list)
}
\arguments{
\item{edge.list}{a 2-column matrix with the
first column being the tails and the second
being the heads.}
}
\value{
a list containing the leaders and sizes
of SCCs.
}
\description{
WeimingSCC::SCC_igraph computes the strongly
connected components using the package igraph.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
